// src/app/features/add-product/models/interfaces/AdPostPayload.interface.ts
export interface AdPostPayload {
  brandId: string;
  modelId: string;
  vehicleTypeId: string;
  bodyTypeId: string;
  engineTypeId: string;
  transmissionTypeId: string;
  drivetrainTypeId: string;
  colorId: string;
  locationTownId: string;
  description: string;
  displacement: number;
  price: number;
  productionYear: number;
  mileage: number;
  vincode: string;
}